# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "commerce-system API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/orders": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "새로운 주문을 생성합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "주문 생성",
                "parameters": [
                    {
                        "description": "주문 정보",
                        "name": "orderRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateOrderRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "주문 생성 성공",
                        "schema": {
                            "$ref": "#/definitions/response.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "주문 생성 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/orders/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "인증된 사용자의 주문 내역을 조회합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "내 주문 조회",
                "responses": {
                    "200": {
                        "description": "주문 목록",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/response.OrderResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "주문 조회 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/orders/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "특정 월의 주문 통계를 조회합니다. (관리자 전용)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "주문 통계 조회",
                "parameters": [
                    {
                        "type": "string",
                        "description": "조회할 월 (YYYY-MM)",
                        "name": "month",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "통계 정보",
                        "schema": {
                            "$ref": "#/definitions/response.OrderStatsResponse"
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "권한 없음",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "통계 조회 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/orders/{order_number}/cancel": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "특정 주문을 취소합니다.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "주문 취소",
                "parameters": [
                    {
                        "type": "string",
                        "description": "주문 번호",
                        "name": "order_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "취소 성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "취소 실패",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "request.CreateOrderRequest": {
            "type": "object",
            "properties": {
                "order_number": {
                    "type": "string"
                },
                "product_number": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "response.OrderResponse": {
            "type": "object",
            "properties": {
                "canceled_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_canceled": {
                    "type": "boolean"
                },
                "member_number": {
                    "type": "string"
                },
                "order_date": {
                    "type": "string"
                },
                "order_number": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "product_number": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                },
                "total_amount": {
                    "type": "integer"
                }
            }
        },
        "response.OrderStatsResponse": {
            "type": "object",
            "properties": {
                "month": {
                    "type": "string"
                },
                "total_canceled": {
                    "type": "integer"
                },
                "total_sales": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "A commerce-system service API in Go using Gin framework",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}